Reduce memory footprint

+++ src/HIDJoystick.cpp
@@ -88,7 +88,7 @@ bool HIDJoystick::parseData(uint8_t *data, uint16_t datalen, HIDJoystickData *jo
                 uint32_t value = HIDUtils::readBitsLE(data, bitOffset, input.size);
                 bitOffset += input.size;
 
-                if (bitOffset > datalen * 8)
+                if (bitOffset > (datalen * (uint32_t)8))
                     return false; // Out of range
 
                 if (input.type == HIDIOType::ReportId)
